fix: refactor and fix minor error
